+++ internal/repository/postgresql/migrations/000006_alter_feeds_table_add_fulltextsearch_column.down.sql
@@ -1,0 +1,7 @@
+-- Copyright (c) VirtualTam
+-- SPDX-License-Identifier: MIT
+
+DROP INDEX IF EXISTS idx_feed_entries_fulltextsearch_tsv;
+
+ALTER TABLE feed_entries
+DROP COLUMN fulltextsearch_tsv;